Pre Save & Post Save Signals
If new object is created then slug field is "this is my slug" otherwise the slug field is slugify(title)

Note: post_save signal contains a extra variable "created: a boolean value" within the definition rest all are same with the pre_save. created is True if new record was created.